[WIP] Refactor Slider
